+++ mlxtend/classifier/stacking_classification.py
@@ -141,10 +141,7 @@ class StackingClassifier(BaseEstimator, ClassifierMixin, TransformerMixin):
 
             for key, value in six.iteritems(super(StackingClassifier,
                                             self).get_params(deep=False)):
-                if key in ('classifiers', 'meta-classifier'):
-                    continue
-                else:
-                    out['%s' % key] = value
+                out['%s' % key] = value
 
             return out