+++ mlxtend/classifier/stacking_cv_classification.py
@@ -245,10 +245,7 @@ class StackingCVClassifier(BaseEstimator, ClassifierMixin, TransformerMixin):
 
             for key, value in six.iteritems(super(StackingCVClassifier,
                                             self).get_params(deep=False)):
-                if key in ('classifiers', 'meta-classifier'):
-                    continue
-                else:
-                    out['%s' % key] = value
+                out['%s' % key] = value
 
             return out